+++ sdk/src/main/java/ly/count/android/sdk/EventQueueProvider.java
@@ -6,5 +6,5 @@ import java.util.Map;
 
 interface EventQueueProvider {
     void recordEventToEventQueue(final @NonNull String key, @Nullable final Map<String, Object> segmentation, final int count, final double sum, final double dur, final long timestamp, final int hour, final int dow, final @NonNull String eventID, final @Nullable String previousViewID,
-        final @Nullable String currentViewId, final @Nullable String previousEventId);
+        final @Nullable String currentViewId, final @Nullable String previousEventId, final @Nullable String previousViewName, final @Nullable String previousEventName);
 }